requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pymupdf==1.22.3
sentence-transformers==2.2.2
scikit-learn==1.2.2
rapidfuzz==2.13.7
python-dateutil==2.8.2
numpy==1.26.4
torch==2.2.3
jinja2==3.1.2
python-multipart==0.0.6
